--- lib/tasks/db.rake
@@ -1,6 +1,0 @@
-namespace :db do
-  desc "Resets the database and loads it from db/dev_seeds.rb"
-  task dev_seed: :environment do
-    load(Rails.root.join("db", "dev_seeds.rb"))
-  end
-end
